Fetch families on row-level instead of table-level - Add DataEntryTableRow - Reorganize DataEntry components - Update useFamiliesQuery to include family_codename param
