+++ react/src/hooks/useFamiliesQuery.tsx
@@ -2,8 +2,8 @@ import { useQuery } from "react-query";
 import { Family } from "../typings";
 import { basicFetch } from "./utils";
 
-async function fetchFamilies() {
-    return await basicFetch("/api/families");
+async function fetchFamilies(familyCodename?: string) {
+    return await basicFetch("/api/families", { starts_with: familyCodename });
 }
 
 /**
@@ -11,7 +11,10 @@ async function fetchFamilies() {
  *
  * That is, return an array of all families.
  */
-export function useFamiliesQuery() {
-    const result = useQuery<Family[], Response>("families", fetchFamilies);
+export function useFamiliesQuery(familyCodename?: string) {
+    const result = useQuery<Family[], Response>(
+        familyCodename ? ["families", familyCodename] : "families",
+        () => fetchFamilies(familyCodename)
+    );
     return result;
 }
